backend/requirements.txt
# ── Core ──
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# ── ML ──
ultralytics>=8.0.200
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# ── TTS ──
gTTS>=2.4.0

# ── Storage ──
boto3>=1.29.0

# ── Admin / Templates ──
jinja2>=3.1.2
itsdangerous>=2.1.2

# ── Data checks / visualization ──
matplotlib>=3.7.0
seaborn>=0.13.0
pandas>=2.0.0
imagehash>=4.3.0

# ── Dev ──
pytest>=7.4.0
httpx>=0.25.0
